Fix tests about invalid distance code
Error can be found where we want to get _base_dist.({30,31}). We should raise
an invalid distance code. Otherwise, we set [d.d] to 0 when we read the extra
distance code. In this case, [dd] silently do the [Write] operation (nothing to
blit). We decide to raise an error in this case, according cover tests of
[zlib].